refactor:(PB-5661) rename favorites services to follow .service.ts naming convention
