+++ src/views/Favorites/services/fetchFavorites.service.ts
