+++ src/views/Favorites/services/toggleFavorite.service.ts
